requirements.txt
django~=2.2.1
requests 
beautifulsoup4
python-telegram-botte
django-bootstrap4